feat: update templates for React 18
