+++ packages/cra-template-wptheme-typescript/package.json
@@ -1,6 +1,6 @@
 {
     "name": "cra-template-wptheme-typescript",
-    "version": "3.4.1-wp.1",
+    "version": "5.0.1-wp.1",
     "keywords": [
         "react",
         "create-react-app",
