+++ packages/cra-template-wptheme-typescript/template.json
@@ -1,14 +1,15 @@
 {
   "package": {
     "dependencies": {
-      "@testing-library/react": "^9.3.2",
-      "@testing-library/jest-dom": "^4.2.4",
-      "@testing-library/user-event": "^7.1.2",
-      "@types/node": "^12.0.0",
-      "@types/react": "^16.9.0",
-      "@types/react-dom": "^16.9.0",
-      "@types/jest": "^24.0.0",
-      "typescript": "~3.7.2"
+      "@testing-library/jest-dom": "^5.16.5",
+      "@testing-library/react": "^13.4.0",
+      "@testing-library/user-event": "^13.5.0",
+      "@types/jest": "^27.5.2",
+      "@types/node": "^16.18.11",
+      "@types/react": "^18.0.27",
+      "@types/react-dom": "^18.0.10",
+      "typescript": "^4.9.5",
+      "web-vitals": "^2.1.4"
     }
   }
 }
