+++ packages/cra-template-wptheme/package.json
@@ -1,6 +1,6 @@
 {
     "name": "cra-template-wptheme",
-    "version": "3.4.1-wp.1",
+    "version": "5.0.1-wp.1",
     "keywords": [
         "wordpress",
         "create-react-app",
