+++ packages/cra-template-wptheme/template/src/App.jsx
@@ -1,1 +1,21 @@
+import React from "react";
+import logo from "./logo.svg";
+import "./App.css";
 
+function App() {
+    return (
+        <div className="App">
+            <header className="App-header">
+                <img src={logo} className="App-logo" alt="logo" />
+                <p>
+                    Edit <code>react-src/src/App.jsx</code> and save to reload.
+                </p>
+                <a className="App-link" href="https://reactjs.org" target="_blank" rel="noopener noreferrer">
+                    Learn React
+                </a>
+            </header>
+        </div>
+    );
+}
+
+export default App;
