+++ packages/create-react-wptheme/package.json
@@ -1,6 +1,6 @@
 {
     "name": "create-react-wptheme",
-    "version": "3.4.1-wp.1",
+    "version": "5.0.1-wp.1",
     "description": "Create React-enabled WP themes.",
     "main": "index.js",
     "scripts": {
